feat: GitHub Action static analysis dependency source check (#20)
* Update comment in dashboard and GitHub Action dependency

Added a comment to clarify code in the dashboard Terraform file. Updated the GitHub Action to reference a new branch for detecting modified folders.

* Add `depends_on` to ensure proper resource provisioning

The `depends_on` attribute was added to link the dashboard creation to the `grafana_folder.synthetic_monitoring` resource. This ensures that the dashboard setup waits for the Grafana folder to be fully provisioned, preventing potential dependency issues.

* Update GitHub Action to use get-modified-folders v1.22.0

Upgraded the action version from v1.21.0 to v1.22.0 in the static analysis workflow. This ensures compatibility with the latest improvements and fixes in the referenced action.