Add `node_os_upgrade_channel` variable and support in AKS
This change introduces the `node_os_upgrade_channel` variable, enabling configuration of the OS upgrade channel for Kubernetes cluster nodes. The variable is optional, with possible values such as Unmanaged, SecurityPatch, NodeImage, and None, defaulting to "None". Updates are made to both the main.tf and variables.tf files to integrate this functionality.